refactor: REINFORCE class inheritance (addressing #1)
