+++ paccmann_generator/__init__.py
@@ -1,4 +1,6 @@
 """Initialize paccmann_generator module."""
-from .reinforce import REINFORCE  # noqa
+from .reinforce_omics import ReinforceOmic  # noqa
+from .reinforce_proteins import ReinforceProtein  # noqa
+
 name = 'paccmann_generator'
 __version__ = '0.0.1'
